feat(sim): density-driven portal discovery + removal (retire neutral EXPLORE)
Portal discovery used to be an agent action (the EXPLORE slider) — but it helps no
faction, so it made a dull, uninteresting slider. Replaced with a neutral, density-
driven SYSTEM process (system/Cycle.managePortalDensity, every checkpoint):

- The portal count converges toward Config.targetPortals() (≈2.5× the onboarding
  startPortals, capped at maxPortals): discovery dominates when sparse (~4:1 near
  empty), evens to ~1:1 at target, removal wins above it. Measured: a board starting
  at 8 grows + stabilises ~13-16 (equilibrium below target because combat also eats
  portals, which discovery replaces). Replaces the old cycle-end removePortals.
- No-space handling: if the map is too packed to place a non-clipping portal
  (Pos.hasPortalSpace, gated by minDistanceBetweenPortals), discovery is skipped and
  that budget rolls into removal — a full board thins out instead of stalling.
- Removed portals shatter + drop mods + destroy all their links/fields (Portal.remove,
  already the case) — confirmed.

EXPLORE retired from QActions + ActionSelector (Explorer.kt deleted) → SliderVector is
now 18 slots (11 QActions + 7 QDestinations); the AI action space is that bit cleaner.
Docs (MECHANICS/ARCHITECTURE/NN/PLAN) updated; also dropped a redundant toDouble.

Recruiting rework + the "leader spends time on neutral actions" idea are NOT in this
commit — flagged for a follow-up (they're bigger design calls).

+++ src/jsMain/kotlin/system/display/Scene3D.kt
@@ -1339,7 +1339,7 @@ object Scene3D {
         // Hide NPCs outside the play area (beyond the inner edge of the boundary wall). Passability isn't
         // computed out there (to save CPU), so NPCs drift/clip along the border — don't render those, nor
         // any inside/walking through the wall. Works for round (inscribed circle) and rectangular fields.
-        if (!Sim.isInPlayArea(npc.pos.x.toDouble(), npc.pos.y.toDouble())) return
+        if (!Sim.isInPlayArea(npc.pos.x, npc.pos.y)) return
         val sphere = Three.Mesh(headGeo, Materials.solid(NEUTRAL_COLOR))
         val gz = groundZ(npc.pos)
         if (Debug.enabled && StuckTracker.isStuck("npc:${npc.id}")) addStuckMarker(sceneX(npc.pos), sceneY(npc.pos), gz)
